backend/requirements.txt
flask==3.0.3
openai==1.51.2
llama-index==0.11.17
pinecone==5.3.1
pinecone-client==5.0.1
flask-cors==5.0.0
requests==2.32.3
fpdf==1.7.2
PyGithub==2.4.0
llama-index-readers-github==0.2.0
together==1.3.3